refactor: moving func to ranker.py
